% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMM_functions.R
\name{combNmatrices}
\alias{combNmatrices}
\title{Combining multiple matrices}
\usage{
combNmatrices(list.matrices, ...)
}
\arguments{
\item{list.matrices}{Lsit of matrices}
}
\value{
Matrix
}
\description{
Liely as independently evolving
}
