% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Discr_maps.R
\name{discr_Simmap}
\alias{discr_Simmap}
\title{Reading unsummarized Stoch Map files from ReVBayes for one tree}
\usage{
discr_Simmap(tree, res)
}
\description{
Reading unsummarized Stoch Map files from ReVBayes for one tree
}
