% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Discr_maps.R
\name{discr_Simmap_all}
\alias{discr_Simmap_all}
\title{Reading unsummarized Stoch Map files from ReVBayes}
\usage{
discr_Simmap_all(tree, res)
}
\description{
Reading unsummarized Stoch Map files from ReVBayes
}
