% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_functions.R
\name{dropDependentTraits}
\alias{dropDependentTraits}
\title{Utility function for cleaning up character data table after amalgamating characters}
\usage{
dropDependentTraits(char_info, dep.mat, td)
}
\description{
Utility function for cleaning up character data table after amalgamating characters
}
