% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_functions.R
\name{filter_coverage}
\alias{filter_coverage}
\title{Utility function for filtering based on missing traits and taxa}
\usage{
filter_coverage(td, traits = 0, taxa = 0)
}
\description{
Utility function for filtering based on missing traits and taxa
}
