% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_functions.R
\name{ontologyHeatMap}
\alias{ontologyHeatMap}
\title{Plots a heatmap along with a phylogeny and trait tree}
\usage{
ontologyHeatMap(td, njt, start = 3, margs = c(0.2, 0.25), ...)
}
\description{
Plots a heatmap along with a phylogeny and trait tree
}
