% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_functions.R
\name{prepareMapsRayDISC}
\alias{prepareMapsRayDISC}
\title{Function for processing revbayes stochastic maps}
\usage{
prepareMapsRayDISC(td, simmaps, discretization_level = 100)
}
\description{
Function for processing revbayes stochastic maps
}
