% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_functions.R
\name{RAC_query}
\alias{RAC_query}
\title{Function for aggregating characters under a specified set of terms (for example, body regions)}
\usage{
RAC_query(char_info, ONT, names)
}
\description{
Function for aggregating characters under a specified set of terms (for example, body regions)
}
