% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_functions.R
\name{writeRevOntologyModels}
\alias{writeRevOntologyModels}
\title{Function to write Ontology CTMC models as a .Rev file for execution in Revbayes}
\usage{
writeRevOntologyModels(td, M, dir, dirW, dirR, dirD)
}
\description{
Function to write Ontology CTMC models as a .Rev file for execution in Revbayes
}
